refactorings and fix libpaths maybe
